Minor Update
Bugfix
Changed location of some methods
